Struct lyon_core::fixed::Fp32

pub struct Fp32<F = _16> {
    // some fields omitted
}

A 32 fixed point number. The size of the fractional is defined by the type parameter F.

Methods

impl<F: FractionalBits> Fp32<F>

fn raw(self) -> i32

Returns the internal representation.

This internal represenataion can be used for computations to avoid bit-shifting between each operation. The number of divisions should be equal to the number of multiplications performed in order to balance out the bit shifts that were skipped.

fn from_raw(bits: i32) -> Self

fn zero() -> Self

fn is_zero(self) -> bool

fn epsilon() -> Self

Smallest increment that can be reresented with this type.

fn from_f32(val: f32) -> Self

Converts from a 32 bits floating point value.

fn to_f32(self) -> f32

Converts to a 32 bits floating point value.

fn from_f64(val: f64) -> Self

Converts from a 64 bits floating point value.

fn to_f64(self) -> f64

Converts to a 64 bits floating point value.

fn sign(self) -> i32

Returns 1 if the number of positive, -1 if it is negative.

fn rem(self, other: Self) -> Self

Returns the result of self % other.

fn min(self, other: Self) -> Self

Returns the lowest of the two values.

fn max(self, other: Self) -> Self

Returns the highest of the two values.

fn abs(self) -> Self

Returns the absolute value of this number.

fn to_fixed<NewF: FractionalBits>(self) -> Fp32<NewF>

Returns the same number with a different fractional precision.

impl<F: FractionalBits> Fp32<F>

fn one() -> Self

fn from_i32(val: i32) -> Self

fn truncate_to_i32(self) -> i32

Convert to an i32, truncating the fractional part

fn mul_div(self, m: Self, d: Self) -> Self

Computes the self * m / d in one go, avoid the precision loss from shifting bits back and forth.

fn min_val() -> Self

fn max_val() -> Self

fn to_fp64<NewF: FractionalBits>(self) -> Fp64<NewF>

Casts into a 64 bits fixed point number.